+++ hotel-backend/src/routes/mod.rs
@@ -17,6 +17,7 @@ pub mod new_inventory;
 pub mod new_invoice;
 pub mod new_maintenance;
 pub mod new_payments;
+pub mod new_products;
 pub mod new_rates;
 pub mod new_reports;
 pub mod new_room_types;